diff: rework node changes
When there is a change on an existing node, the returned objects are not
consistent in API terms. Some are "xxxChanged" and some are "xxxAdded"
or "xxxRemoved". Moreover, all of these objects have different
attributes and methods. This makes it hard to exploit the results.

Rework the objects so that there are only "xxxAdded" and "xxxRemoved"
objects. Also, make sure that all of these objects inherit from the same
parent class so that we have a consistent API.

Signed-off-by: Robin Jarry <robin@jarry.cc>