setup: do not link with a static libyang
When libyang is built by us, we need to make it available to other CFFI
bindings (such as sysrepo).

Also, this makes static building a problem since we need a single .so
lib to allow proper interaction with libsysrepo.so.

Disable the static build, install the compiled libyang.so into the
libyang package (into a _lib dir) along with libyang extensions and user
types plugins. Force a relative runtime path on the CFFI binding so that
it finds libyang.so in that _lib dir. Since libyang needs to load
plugins with dlopen and the plugin default paths are hardcoded during
the build, force environment variables *before* loading the _libyang
extension.

Also install the C headers in the libyang package (in a _include dir)
for other CFFI bindings (such as sysrepo) to use.

Add functions to get the lib dirs and include dirs to use when building
other CFFI bindings.

Installing in a virtualenv should work out of the box.

Signed-off-by: Robin Jarry <robin@jarry.cc>