Remove LIBYANG_INSTALL=embed support
Actually embedding libyang.so into the python package requires to
maintain the C library code as a git submodule and manually updating it
from time to time. Also, this implies a strong dependency between the
two projects where only a SO version dependency is actually required.

For the sake of simplicity and to allow more flexibility in libyang
python versions:

- Remove the C library code from this repo.
- Remove LIBYANG_INSTALL=embed support.
- Change tox environments accordingly.
- Update README accordingly.
- Check for libyang SO version while compiling the CFFI module.

Signed-off-by: Robin Jarry <robin@jarry.cc>