data: use name cache for DNode.print_dict
When converting a large number of list elements, the conversion of the
lyd_node->schema->name ("const char *" to python "str") takes
a significant time and the same operation is done once per list element.

Store the python string in a cache indexed by the schema node to avoid
doing it more than once.

This reduces the time spent by ~20% for converting a large number of
elements of a list.

Signed-off-by: Robin Jarry <robin@jarry.cc>