schema: add state method for conditional if-features
Accessing feature states is useful to enable/disable leafs.
However, when a if-feature contain a condition (not/or/and keywords),
there is currently no way to get the state resulting of the expression.

Add a state method to all IfFeatureExpr objects so a state of the
entire if-feature expression can be evaluated.

Test all basic cases (one operator).