context: simplify parse flag names
The flag arguments names are very long and this makes hard to read code.

Remove the parser_ and validation_ prefixes to the argument names. There
are almost no name collisions except for no_state. Since it makes no sense
to pass this flag only to validation or parser, it translates to both
validation and parser.