context: fix duplicate path error
search_dirs can contain duplicate path if YANGPATH envvar is used.
With libyang1, we can add several time the same searchdir without
problem but with libyang2, ly_ctx_set_searchdir returns an error if we
add a path already existant.

To fix it, we now pass the search dir directly when creating the
context.