dnode: add iter_tree
The DNode.diff function result is hard to use because it is not possible
to iterate on the node tree. Add DNode.iter_tree for this purpose. The
code mimics the LYD_TREE_DFS_BEGIN/LYD_TREE_DFS_END libyang macros, but
in python.

Another possibility was to use the macros directly and call a python
callback, but it is easier to maintain this way.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>