schema: fix type patterns return value
The libyang lysp_type struct arg.str fields contains the pattern string
+ a character in the beginning telling if the match is inverted or not.

Since the libyang2 port, the patterns function do not take this into
account anymore, and yields this string directly, which is not correct.

The patterns function is supposed to yield Iterator[Tuple[str, bool],
containing the pattern string and a boolean telling if the match is
inverted or not, restore it. Add a test for it as well.

While there, use ly_array_iter.

Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>