schema: return float for decimal64 in default function
This patch changes the output for leaf/leaf-list nodes using decimal64
type, and adds new unit test for this purpose.

The leaf-list defaults function is reworked to be similar to the leaf
default function. A new yang schema is added for clarity, and to avoid
modifying all the tests using yolo-system.yang.

Fixes: #80
Signed-off-by: Stefan Gula <steweg@gmail.com>
Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>