data: fix DNode.new schema handling
In case of a call to DNode.new with cdata containing an opaque node
the new method tries to access cdata.schema.nodetype, which results
in a NULL pointer dereference.

To get the schema for an opaque node retrieve the schema from the
context using the path of the node.

Fixes: #73
Signed-off-by: nvxf <68589039+nvxf@users.noreply.github.com>