schema: support LYS_GETNEXT_WITH{CASE,CHOICE}
Add support for LYS_GETNEXT_WITH{CASE,CHOICE} for Module, Container,
Choice and Case. To do so, introduce a new iter_children_options
function that translates booleans to lys_getnext options.

The Rpc children function is updated to use the iter_children_options
mecanism. The lib.LYS_CHOICE and lib.LYS_CASE are added by to the
default types in iter_children, as it does not change the behavior.

Fixes: #91
Signed-off-by: Stefan Gula <steweg@gmail.com>
Signed-off-by: Samuel Gauthier <samuel.gauthier@6wind.com>