xpath: add support for indexed xpath
This patch adds support for indexed xpath to xpath_set.
In the case of a leaf-list, it is now possible to apply xpath_set in this way:

xpath_set(d, "/lstnum[5]", 33, after="4")

This will replace the 5th element of the leaf-list with the new value 33.

This feature is important because libyang can return this type of xpath
when a callback is called with sr_module_change_subscribe from sysrepo.

The tests are updated accordingly.

Signed-off-by: Jean-Sébastien Bevilacqua <jean-sebastien.bevilacqua@6wind.com>