context/data: memleak due to API (mis)use
Memory management is a foreign concept to python in general, however it
appears DNode and Context require .free() and .destroy() to be called,
respectively, to avoid a memory leak.

This seems easily avoidable by simply attaching a destructor to the
class so the python garbage collector can clean up automatically when all
references go out of scope.

We add a reference to any DNode object derived from another DNode object
(but not duplicated) so we can determine if we are allowed to be the one
to destroy the DNode cdata, as well as to keep proper reference counting
in case the owner goes out of scope but a dependent reference is still
in scope.

This change should also be compatible with existing integrations which may be
aware of this as it sets the internal `cdata` reference to None and checks
it to ensure it doesn't run the cleanup again.  That said, I think calling
DNode.free() in this case is still risky since there are no checks to
ensure this is truly valid like the destructor does.

Signed-off-by: Brad House <brad@brad-house.com>