log: expose ly_temp_log_options with a context manager
When one wishes to disable logging temporarily (e.g., when calling API functions
when an error result is expected and OK), libyang provides
ly_temp_log_options(). We need access to this in python, so export access to the
function and add a python context (i.e., "with") manager API for using
it idiomatically as well.

ex usage:

```
    with temp_log_options(0):
        ly_unwanted_logging_call();
```

Signed-off-by: Christian Hopps <chopps@labn.net>