tests: test_iffeature_state crash fix
As observed on Debian Trixie RC2, test_iffeature_state crashes due
to an internal pointer being invalidated.  This invalidation appears
to be due to a call to lys_set_implemented() possibly causing a
full recompile of the ctx.

This simply reorders the caching of the path pointers so they are
not invalidated when used.

Signed-off-by: Brad House <brad@brad-house.com>